+++ cogs/admin.py
@@ -3,14 +3,12 @@ import discord
 from discord.ext import commands
 from utils.logroll import logging
 log = logging.getLogger(__name__)
-from utils.pgsql import (
-    drop_all_tables,
-    create_tables,
-    set_reply_channel,
-    drop_reply_channel,
-    drop_user_input_ids,
-    drop_guild_input_ids,
-    set_reply_to
+from utils.nosj import (
+    drop_guild_input_tokens,
+    set_key,
+    reset_file_data,
+    save_file_data,
+    show_file_data
 )
 
 class Admin(commands.Cog):
@@ -26,28 +24,35 @@ class Admin(commands.Cog):
         if not user:
             await ctx.send("Forget who? I need a username")
             return
-        drop_user_input_ids(user.id)
+        await set_key(ctx.guild.id, user.id, "input_tokens", [])
         await ctx.send(f"I have no memory of {user.mention}!")
 
     @commands.command()
     @commands.has_permissions(administrator=True)
     async def forgetguild(self, ctx):
         """Clear all chat memory of this Guild."""
-        drop_guild_input_ids(ctx.guild.id)
+        await drop_guild_input_tokens(ctx.guild.id)
         await ctx.send(f"I have no memory of {ctx.guild.name}.")
+
+    @commands.command()
+    @commands.has_permissions(administrator=True)
+    async def resetguild(self, ctx):
+        """Clear all chat memory of this Guild."""
+        await drop_guild(ctx.guild.id)
+        await ctx.send(f"Guild {ctx.guild.name} data reset.")
 
     @commands.command()
     @commands.has_permissions(administrator=True)
     async def setreplychannel(self, ctx):
         """Set moe's designated reply channel."""
-        set_reply_channel(ctx.guild.id, ctx.channel.id)
+        await set_key(ctx.guild.id, None, "reply_channel", ctx.channel.id)
         await ctx.send(f"Reply channel set to {ctx.channel.mention}")
 
     @commands.command()
     @commands.has_permissions(administrator=True)
     async def unsetreplychannel(self, ctx):
         """Unset the reply channel for the guild."""
-        drop_reply_channel(ctx.guild.id)
+        await set_key(ctx.guild.id, None, "reply_channel", None)
         await ctx.send("Reply channel unset.")
 
     @commands.command()
@@ -57,7 +62,7 @@ class Admin(commands.Cog):
         if not user:
             await ctx.send("I'm not a mind-reader you know, tell me a name!")
             return
-        set_reply_to(ctx.guild.id, user.id, True)
+        await set_key(ctx.guild.id, user.id, "reply_to", True)
         await ctx.send(f"Hello {user.name}!")
 
     @commands.command()
@@ -67,12 +72,13 @@ class Admin(commands.Cog):
         if not user:
             await ctx.send("Bye to who?")
             return
-        set_reply_to(ctx.guild.id, user.id, False)
+        await set_key(ctx.guild.id, user.id, "reply_to", False)
         await ctx.send(f"Bye {user.name}!")
 
-    @commands.command(name="reload", hidden=True)
+    @commands.command()
     @commands.is_owner()
     async def reload(self, ctx, extension: str):
+        """Reload a module"""
         if not extension:
             await ctx.send("Reload which module? Be specific.")
             return
@@ -84,13 +90,33 @@ class Admin(commands.Cog):
             log.error(f"Failed to reload cogs.{extension}, invoked by {ctx.author}")
             await ctx.send(f"Failed to reload `{extension}`: {e}")
 
-    @commands.command(name="rebuild", hidden=True)
+    @commands.command()
     @commands.is_owner()
     async def rebuild(self, ctx):
+        """Wipe and rebuild the memory file."""
         log.warning(f"{ctx.author} invoked database rebuild")
-        drop_all_tables()
-        create_tables()
+        await reset_file_data()
         await ctx.send(f"Rebuild complete!")
+
+    @commands.command()
+    @commands.is_owner()
+    async def shutdown(self, ctx):
+        """Shut the bot down"""
+        log.info("Shutdown command received")
+        await ctx.send("Nooo...")
+        await save_file_data()
+        await self.bot.close()
+
+    @commands.command()
+    @commands.is_owner()
+    async def dumpbackend(self, ctx):
+        """Dump memory file to logs"""
+        try:
+            json_data = await show_file_data()
+            log.info(f"{json_data}")
+            await ctx.send("Dumped backend logs! lol")
+        except Exception as e:
+            await ctx.send(f"Could not dump! {e}")
 
 async def setup(bot):
     await bot.add_cog(Admin(bot))