+++ cogs/tools.py
@@ -1,13 +1,16 @@
 # cogs/tools.py
 import discord
 from discord.ext import commands
-from utils.pgsql import drop_user_input_ids
+from utils.dialogpt import dialogpt
+from utils.nosj import (
+  get_input_tokens,
+  set_key
+)
 
 class Tools(commands.Cog):
-    """User accessible commands."""
-
     def __init__(self, bot):
         self.bot = bot
+        self.tokenizer = dialogpt.load_tokenizer()
 
     @commands.command()
     @commands.has_permissions(send_messages=True)
@@ -19,8 +22,24 @@ class Tools(commands.Cog):
     @commands.has_permissions(send_messages=True)
     async def forgetme(self, ctx):
         """Erase your chat history from møe's memory"""
-        drop_user_input_ids(ctx.author.id)
-        await ctx.send(f"I have no memory of {ctx.author.mention}!")
+        try:
+            await set_key(ctx.guild.id, ctx.author.id, "input_tokens", [])
+            await ctx.send(f"I have no memory of {ctx.author.mention}!")
+        except Exception as e:
+            await log.warning(f"Could forget user: {e}")
+
+    @commands.command()
+    @commands.has_permissions(send_messages=True)
+    async def chatlogs(self, ctx):
+        """Show your chat history with møe from this Guild"""
+        try:
+            chat_history = await get_input_tokens(ctx.guild.id, ctx.author.id)
+            chat_history = self.tokenizer.decode(chat_history, skip_special_tokens=False).split("<|endoftext|>")
+            chat_history = "\n".join(message.strip() for message in chat_history if message.strip())
+            await ctx.send(f"{ctx.author}'s chat history: \n```{chat_history}```")
+        except Exception as e:
+            await ctx.send(f"I've never spoken to {ctx.author}!")
+            await log.warning(f"Could not get user history: {e}")
 
 async def setup(bot):
     await bot.add_cog(Tools(bot))