+++ utils/conf.py
@@ -1,37 +1,35 @@
 # utils/conf.py
 import os
 from py_dotenv import read_dotenv
-dotenv_path = os.path.join(os.path.dirname(__file__), '..', '.env')
+dotenv_path = os.path.join(os.path.dirname(__file__), "..", ".env")
 read_dotenv(dotenv_path)
-
-pgsql_db = os.getenv('PGSQL_DB')
 
 class Config:
     def __init__(self):
-        self.discord_token   = os.getenv('AUTH_TOKEN')
-        self.discord_prefix  = os.getenv('CMD_PREFIX') + ' '
+        self.discord_token   = os.getenv("AUTH_TOKEN")
+        self.discord_prefix  = os.getenv("CMD_PREFIX") + " "
         self.discord_intents = self.get_discord_intents()
         self.reaction_emojis = self.get_reaction_emojis()
 
     @staticmethod
     def get_discord_intents():
         return {
-            'guilds': True,
-            'members': True,
-            'messages': True,
-            'reactions': True,
-            'message_content': True,
+            "guilds": True,
+            "members": True,
+            "messages": True,
+            "reactions": True,
+            "message_content": True,
         }
 
     @staticmethod
     def get_reaction_emojis():
         return {
-            'upvote': os.getenv('UPVOTE_EMOJI'),
-            'downvote': os.getenv('DOWNVOTE_EMOJI'),
-            'noyou': os.getenv('NOYOU_EMOJI'),
-            'think': os.getenv('THINK_EMOJI'),
-            'moeji': os.getenv('MOE_EMOJI'),
-            'wow': os.getenv('WOW_EMOJI'),
+            "upvote": os.getenv("UPVOTE_EMOJI"),
+            "downvote": os.getenv("DOWNVOTE_EMOJI"),
+            "noyou": os.getenv("NOYOU_EMOJI"),
+            "think": os.getenv("THINK_EMOJI"),
+            "moeji": os.getenv("MOE_EMOJI"),
+            "wow": os.getenv("WOW_EMOJI"),
         }
 
 config = Config()